pyproject.toml
[project]
name = "water-ontology"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "pyontoenv==0.4.0a4",
    "rdflib>=7.0.0",
    "buildingmotif[topquadrant]",
    "brick-tq-shacl==0.4.0a7",
    "jupyterlab>=4.2.5",
    "jupyterlab-rise>=0.42.0",
    "ipykernel>=6.29.5",
    "jupyter-book>=1.0.3",
    "sphinx-autodoc-bmotif==0.1.6",
    "sphinx-toolbox>=3.8.1",
    "sphinx-tabs>=3.4.5",
    "networkx>=3.4.2",
    "pydotplus>=2.0.2",
    "graphviz>=0.21",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv.sources]
buildingmotif = { git = "https://github.com/NREL/buildingmotif.git", rev = "develop" }

[tool.uv]
override-dependencies = ["brick-tq-shacl==0.4.0"]
